requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
JPype1==1.5.0
optapy==9.37.0b0
packaging==24.0
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse @ file:///C:/b/abs_88361ub_qu/croot/sqlparse_1690904577514/work
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
